Refactor tests to separate remote and non-remote methods. Add setup_package.py to include parameter_list.json during build phase.
