+++ astroquery/dace/tests/setup_package.py
@@ -1,0 +1,9 @@
+# Licensed under a 3-clause BSD style license - see LICENSE.rst
+from __future__ import absolute_import
+
+import os
+
+
+def get_package_data():
+    paths = [os.path.join('data', '*.json')]
+    return {'astroquery.dace.tests': paths}
